ADDITION OF FORCE_FORK PARAMETER FOR PROXIES
Proxy.pm:

Handle the force_fork parameter.  If it is 1 then when we get
a connection from a client we fork, if it is 0 then we handle
the connection in line and queue up connections using the standard
TCP mechanism up to the maximum backlog for the platform.


POP3.pm:
NNTP.pm:

Default to not forking


SMTP.pm:

Default to forking


git-svn-id: file:///usr/local/svn/trunk@1267 215d8e1a-9f62-4c3a-9943-246481b8efb3