BROUGHT TEST SUITE TO 100% AFTER BERKELEYDB CHANGE
HTML.pm:

Small modification to the way that showing an individual bucket works
so that when it is cleared we show the same page with a zero word
count and no words.

Bayes.pm:

Mechanism for marking a bucket deleted even though the files in
its directory have not all disappeared.  This is needed because if
a user deletes a bucket while a download is occurring POPFile will
have a connection open to the db which will prevent it from getting
deleted.  Now have a helper delete_bucket_files__ that will do the
delete and add the special delete file as needed.

clear_bucket can't rely on being able to remove the table.db for
the same reasons as the above, so change the way in which it operates.

TestBayes.tst:

Tests to cover upgrading of corpus tables, fork handling and deal
with the fact that the corpus is no longer in flat files.

TestHTML.tst:

Have to call prefork/forked/postfork when forking.  Test the new
force_fork UI elements.

TestPOP3.tst:

Have to call prefork/forked/postfork when forking.

TestProxy.tst:

Add code to test the forking/non-forking version of the proxy code.


git-svn-id: file:///usr/local/svn/trunk@1289 215d8e1a-9f62-4c3a-9943-246481b8efb3