Fix bug 826765
@ and $ inside magnets were not being handled properly.

Classifer/Bayes.pm:

Factor most of magnet_match__ into magnet_match_helper__ so
that there is no duplicated code.  Remove use of regexps for
magnet match and replace with simple 'eq' matching, thus
eliminating all the complexities around special characters
in regexps and the fact that @ and $ are illegal in \Q \E
quoted regexps.

tests/TestBayes.tst:

Added tests for magnet_match__ with specific emphasis on
handling of $ and @.

Made Japanese tests detect whether Text::Kakasi is present
on the machine and ignore them (with a warning if it is
not present).


git-svn-id: file:///usr/local/svn/trunk@1420 215d8e1a-9f62-4c3a-9943-246481b8efb3