Make test suite pass 100%.   Make POPFile cope with illegal CRLF.CRLF in the middle of a mail message.  Fix an oddity in our TOP caching.
NOTE: The HTML test suite is still a little flaky.  It sometimes dies
quickly complaining about an illegal value 'John' for input 'text2'

Makefile
tests/Makefile:

Add new 'eachtest' target that runs each test in a for loop to
help with some test debugging.

Makefile:

New 'killperls' target that is used (on Linux) to kill any Perl processes
that are still running, this target is used when debugging the test suite
and the test suite is leaving perl subprocesses alive.

Classifier/Bayes.pm:

Corpus upgrade messages now go to the log and not STDOUT.

Strip the X-POPFile-Link header from incoming mail, as well as the
X-Text-Classificationheader.  Remove any reference to the 'unsure'
classification since we are not using it.

Call flush_extra_ after getting a message, in case there is data
after the first CRLF.CRLF in the message.  This should be illegal but
we have an example of a real mail server sending messages like this
which was causing POPFile to fail.

Don't specify an ErrFile for the BerkeleyDB so that we don't get lots
of error files generated.

echo_to_dot_ now returns 0 if it doesn't find the . before EOF

Use new echo_to_dot_ return code to handle an odd situation where if we
were doing classify_and_modify from a file (which didn't end with . ) and
the message was very large (> 100,000 bytes) it was possible that the
function didn't send a . to the client handle.

In the QuickMagnets code make each <option> have a specific value, instead
of assuming that the value is the same as the text in the <option>.

POPFile/Logger.pm:

The location of the log files is relative to the user root.

POPFile/Module.pm:

Fix a bug in flush_extra_ that was causing a failure in the POP3 module, it
seems that if you do 'last;' in anywhere it exits the nearest enclosing
loop EVEN IF THAT LOOP IS NOT IN THE SAME FUNCTION.  e.g.

while (1) {
    function();
}

will terminate if function() does a last; outside a loop.

Proxy/POP3.pm:

In the RETR after TOP code we were sending .$eol to the client because of
the bug in Bayes.pm mentioned above, but this was errorneous too because it
would then be possible for us to send .$eol.$eol which would have been
happening for all short messages using RETR after TOP.

Fix bug in code that shortens a path to just a file name, if the path
contained multiple / then it didn't work.

POPFile/Configuration.tst:

Set up the POPFILE_USER and POPFILE_ROOT during new() so that they are valid
in all PLM's initialize() functions.

tests/TestPOP3.tst:

Added tests using TestMailParse028 to check for correct operation when an
illegal CRLF.CRLF is encountered in a message.

Close the server handle on the fake POP3 server, this was causing tests to
be unreliable on Linux.

tests/TestBayes.tst:

Update tests for the new return code from echo_to_dot_.

tests/TestConfiguration.tst:

Update tests to handle new stripping of invalid parameters from the
cfg file.  Test a valid new parameter, and the stripping of a parameter.

tests/TestMailParse012.cam
tests/TestMailParse018.cam
tests/TestMailParse019.cam:

Since XPL is now correctly stripped, need to update these tests as they
have an XPL in the MSG file.

tests/TestMailParse028.msg
tests/TestMailParse028.cls
tests/TestMailParse028.wrd
tests/TestMailParse028.cam:

Added test cases where there's an embedded CRLF.CRLF in the middle of the
message.

tests/TestHTML.script:

Fix a few tests that failed because Stanley fixed an HTML 4.01 compliance bug
where we were using & and it should have been &amp;  Minor fix since we
changed the way we should exponentation (e-005 is now e-05).  Change the tests
for loading various files since they are either relative to the POPFILE_ROOT
or POPFILE_USER.

Small change to delete test since we've added extra messages to the tests.

tests/TestPipeScript.tst:

Remove any popfile.cfg that's hanging around so that these tests use defaults.

tests/corpus.base/spam/table
tests/corpus.base/other/table
tests/corpus.base/personal/table:

Convert line endings to just LF and remove blank lines.


git-svn-id: file:///usr/local/svn/trunk@1508 215d8e1a-9f62-4c3a-9943-246481b8efb3