Merged David's patch to have a separate connection for each serviced folder.
Corrected a stupid little bug that led to unsorted folder lists in the UI

Added a check for buckets that don't have a corresponding folder set, so
we won't see any warnings or attempts to SELECT ''.

Changed how the IMAP tag is handled. It is now initialized to 0 in new()
and is no longer reset in login. Thus all connections can have unique tags
and all tags will be unique for the complete module. By the time the tag
reaches 9999, command 0 should have been processed ;-)


git-svn-id: file:///usr/local/svn/trunk@1907 215d8e1a-9f62-4c3a-9943-246481b8efb3