Added a global flag that tells us when the list of folders has changed. This can happen when the user adds watched folders via the UI or when the user changes the bucket to folder mapping.
Revamped the connection pool. The $self->{imap__} hash no longer
exists. Instead, connections are kept along with other per-folder
information in $self->{folders__} as in $self->{folder__}{"INBOX"}{imap}.

The list of folders in $self->{folder__} is now compiled in a new
function: build_folder_list__()

connect_folders__() now honors pre-existing connections and handles
failed connects, LOGINs, and SELECTs more gracefully.

The UIDVALIDITY value for a folder is no longer retrieved with an
extra STATUS commmand. Instead it is extracted from the server's
response to the SELECT commmand. A STATUS is now only done for folders
we access for the very first time to find out about UIDNEXT.

Changed the hash function to be compatible to what the history module
will do. This means that the hashes of the messages in your history
will no longer correspond to the hashes computed when IMAP.pm encounters
new messages. To be completely on the safe side, purge your current
history before updating to this version of IMAP.pm

Implemented first steps towards handling of unsolicited messages from
the server. To this end, the last command sent to the server is stored
in $self->{last_command__} and in get response we look at the command
and decide whether any untagged responses are solicited or not.
Currently, log entries are generated for unsolicited BYE and UIDVALIDITY
responses.


git-svn-id: file:///usr/local/svn/trunk@1912 215d8e1a-9f62-4c3a-9943-246481b8efb3