Changed connect_folders__() to not connect to mailboxes that are write-only (output folder that are not watched and that receive the messages classified to pseudo-buckets).
Fixed bug in folder_uid_status__ that prevented to store the new
UIDNEXT value in case of a changed UIDVALIDITY.

Fixed bug in get_new_message_list() that returned an alphabetically
sorted list of UIDs instead of a numerically sorted list.

Removed debug-only log_() statements that kept an idea on the
interaction with the history module.

Other very minor changes.


git-svn-id: file:///usr/local/svn/trunk@1985 215d8e1a-9f62-4c3a-9943-246481b8efb3