Changed test suite to be more tool (and therefore platform) independent:
test.pl now 'use's File::Copy, File::Find, and File::Path;

These provide for:
copy( src, dst), find(), and rmtree(dir), respectively.

Find is used by the new sub in tests.pl, rec_cp( src, dst ), which will copy a directory tree recursively.

There is still some work to do in this regard. The biggest headache is the shell function 'date'. Even if one gets the GNU date program, the MS shell will not use it and prefer its built-in date routine. 'date' is used in TestHTML.tst and TestLogger.tst. Of course, a wrapper around the call to date could be made that avoids the nifty GNU date command line options and boils it down to a call that even MS date will understand.

There are also some calls to rm in the makefile.


git-svn-id: file:///usr/local/svn/trunk@2006 215d8e1a-9f62-4c3a-9943-246481b8efb3