This is the first version of the tests for IMAP.pm.
IMAP.pm got a new function disconnect_folders__() that will disconnect so that the test module can change the IMAP settings and start a new round of testing.

IMAP_test_server.pl is what the name suggest, a test server. It mimics an IMAP server, but it's tailored to the conventions used in IMAP.pm. For example, IMAP.pm will always quote the name of folders and so the server will refuse to deal with unquoted folder names. Another example are the tags that need to prefix each IMAP command. The server expects tags that conform to the rx /^A\\d+/.

testIMAP.tst forks itself into a parent that will do all the testing and a single child that runs the server.

Current test coverage for IMAP.pm: 38%.


git-svn-id: file:///usr/local/svn/trunk@2065 215d8e1a-9f62-4c3a-9943-246481b8efb3