Change the way in which the "Browse" button works. The NSIS default may append "\POPFile" to the selected path which can cause confusion, e.g. the user selects "C:\Program Files\Test" but the installer installs the program in "C:\Program Files\Test\POPFile". Although the DIRECTORY page displays "C:\Program Files\Test\POPFile" most users will click the 'Next' button without noticing the extra "\POPFile" at the end of the path.
If the user types a SFN path in the DIRECTORY page, try to convert this
to a LFN path before using it in the installer wizard. Stop accepting
$PROGRAMFILES as the installation folder (e.g. if $PROGRAMFILES is
"C:\Program Files" then do not allow "C:\Program Files" to be selected
as the installation folder for POPFILE_ROOT or POPFILE_USER).


git-svn-id: file:///usr/local/svn/trunk@2386 215d8e1a-9f62-4c3a-9943-246481b8efb3