More work on v0.23.0 (Multi-user Support)
Add the 'Users' tab that is available only to administrators and
allows them to create, remove and edit users.  Users can be created as
a clone of an existing user and the complete list of user parameters
can be edited by the administrator.

There is currently no way to use this information in POPFile (that's
for my next commit); POPFile is still hard-wired to use any
information related to user 1 (admin).

----

Classifier/Bayes.pm:

Make the cache of user parameter values have entries for whether the
value is default or not and also keep track of the last time that the
cache entry was read or written so that later we can implementing
culling of the cache to save space.

New APIs added that are accessible by authenticated administrators:
create_user (with optional cloning of another user), remove_user,
get_user_list (returns a list of all users), get_user_parameter_list
(returns the list of all parameters a user can have), get_user_id
(returns the database ID for a named user).

Add get_user_parameter as way to get a user parameter for the
authenticated user.

Change signature of get_user_parameter_from_id so that it returns both
the value and whether the value is the default or not.

POPFile/Module.pm:

Update user_module_config_ because of the change in the signature of
get_user_parameter_from_id.

UI/HTML.pm:

Add support for the users_page (using the users-page.thtml template) and handle create, remove and edit of users.

New functions status_message__ and error_message__ are used to set a
status or error message using the common-middle.thtml bars at the top
of the UI.  The intention is that all status and error messages are
displayed in a standard place.

languages/English.msg:

Add strings needed by the new Users page.

skins/default/style.css:

New styles shellStatusMessage and shellErrorMessage for styling the
display of status and error messages.  The HTML for this is in
common-middle.thtml.

Remove DOS-style line endings.

skins/default/common-middle.thtml:

Add bars at the top of the screen above the tabs that can be used to
display status and error messages.

skins/default/advanced-page.thtml:

Left justify the column headers on the list of user parameters because
it looks better.

skins/default/history-bar.thtml:

Remove error messages that are now handled in HTML.pm by calling
error_message__ with the appropriate error string.

skins/default/users-page.thtml:

Add HTML to handle the Users page.

----

TODO: Go through code and make all errors and status message use the
new status_message__ and error_message__ functions in HTML.
TODO: Cache flushing of user parameters in Bayes.
TODO: New Bayes API functions need tests


git-svn-id: file:///usr/local/svn/trunk@2461 215d8e1a-9f62-4c3a-9943-246481b8efb3